feat: Add support for path-style S3 URLs
Adds a new configuration option `UsePathStyle` to the S3 client
that allows using path-style URLs instead of the default
virtual-hosted-style URLs. This is useful for S3-compatible
storage services that do not support virtual-hosted-style URLs.

The new option is controlled by the `USE_PATH_STYLE` environment
variable, which can be set to `"true"` to enable path-style URLs.